scripts/requirements.txt
plotly>=5.18.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
